renamed and added additional info
